+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Fix an error for `RSpec/Extra/RestrictBlockTag`. ([@ydah])
+- Fix an error for `RSpec/Extra/RestrictBlockTagValue`. ([@ydah])
 
 ## 0.1.0 - 2023-04-18